older regridmixing send to heaven
